Do not push or pull translations on pull_request event
Allow to run tests without wasting API requests

+++ locale/update.sh
@@ -14,8 +14,13 @@ cd `dirname $0`
 sphinx-build -T -b gettext ../sphinx/doc pot
 sphinx-intl update-txconfig-resources -p pot -d .
 cat .tx/config
-tx push -s --skip
-rm -R -f ar ca_ES zh_CN fr de it_IT ja ko pl_PL pt_BR ru sr es
-tx pull --silent -f -l ar,ca_ES,zh_CN,fr,de,it_IT,ja,ko,pl_PL,pt_BR,ru,sr,es
+
+# Skip Transifex commands when running pull requests event on GitHub Actions
+if [[ -z "$CI" || "$GITHUB_EVENT_NAME" != 'pull_request' ]]; then
+  tx push -s --skip
+  rm -R -f ar ca_ES zh_CN fr de it_IT ja ko pl_PL pt_BR ru sr es
+  tx pull --silent -f -l ar,ca_ES,zh_CN,fr,de,it_IT,ja,ko,pl_PL,pt_BR,ru,sr,es
+fi
+
 git checkout .tx/config
